+++ src/spring-cloud/azext_spring_cloud/commands.py
@@ -50,7 +50,7 @@ def load_command_table(self, _):
         g.custom_command('repo update', 'config_repo_update')
         g.custom_command('repo list', 'config_repo_list')
 
-    with self.command_group('spring-cloud app', client_factory=cf_spring_cloud_20210601preview,
+    with self.command_group('spring-cloud app', client_factory=cf_spring_cloud_20210901preview,
                             exception_handler=handle_asc_exception) as g:
         g.custom_command('create', 'app_create')
         g.custom_command('update', 'app_update')
@@ -71,6 +71,7 @@ def load_command_table(self, _):
         g.custom_command('restart', 'app_restart', supports_no_wait=True)
         g.custom_command('logs', 'app_tail_log')
         g.custom_command('append-persistent-storage', 'app_append_persistent_storage')
+        g.custom_command('append-loaded-public-certificate', 'app_append_loaded_public_certificate')
 
     with self.command_group('spring-cloud app identity', client_factory=cf_spring_cloud,
                             exception_handler=handle_asc_exception) as g:
@@ -113,12 +114,13 @@ def load_command_table(self, _):
         g.custom_command('remove', 'storage_remove')
         g.custom_command('list-persistent-storage', "storage_list_persistent_storage", table_transformer=transform_app_table_output)
 
-    with self.command_group('spring-cloud certificate', client_factory=cf_spring_cloud,
+    with self.command_group('spring-cloud certificate', client_factory=cf_spring_cloud_20210901preview,
                             exception_handler=handle_asc_exception) as g:
         g.custom_command('add', 'certificate_add')
         g.custom_show_command('show', 'certificate_show', table_transformer=transform_spring_cloud_certificate_output)
         g.custom_command('list', 'certificate_list', table_transformer=transform_spring_cloud_certificate_output)
         g.custom_command('remove', 'certificate_remove')
+        g.custom_command('list-reference-app', 'certificate_list_reference_app', table_transformer=transform_app_table_output)
 
     with self.command_group('spring-cloud app custom-domain', client_factory=cf_spring_cloud,
                             exception_handler=handle_asc_exception) as g: