+++ src/spring-cloud/azext_spring_cloud/custom.py
@@ -6,12 +6,13 @@
 # pylint: disable=unused-argument, logging-format-interpolation, protected-access, wrong-import-order, too-many-lines
 import requests
 import re
+import os
 
 from azure.core.exceptions import HttpResponseError
 from azure.mgmt.cosmosdb import CosmosDBManagementClient
 from azure.mgmt.redis import RedisManagementClient
 from requests.auth import HTTPBasicAuth
-import yaml   # pylint: disable=import-error
+import yaml  # pylint: disable=import-error
 from time import sleep
 from ._stream_utils import stream_logs
 from msrestazure.tools import parse_resource_id, is_valid_resource_id
@@ -30,7 +31,7 @@ from .vendored_sdks.appplatform.v2021_09_01_preview import (
 )
 from knack.log import get_logger
 from .azure_storage_file import FileService
-from azure.cli.core.azclierror import InvalidArgumentValueError, RequiredArgumentMissingError
+from azure.cli.core.azclierror import ClientRequestError, FileOperationError, InvalidArgumentValueError, RequiredArgumentMissingError
 from azure.cli.core.commands.client_factory import get_mgmt_service_client
 from azure.cli.core.util import get_file_json, sdk_no_wait
 from azure.cli.core.profiles import ResourceType, get_sdk
@@ -45,6 +46,7 @@ from threading import Thread
 from threading import Timer
 import sys
 import json
+import base64
 from collections import defaultdict
 
 logger = get_logger(__name__)
@@ -246,7 +248,8 @@ def list_keys(cmd, client, resource_group, name, app=None, deployment=None):
 
 # pylint: disable=redefined-builtin
 def regenerate_keys(cmd, client, resource_group, name, type):
-    return client.services.regenerate_test_key(resource_group, name, models.RegenerateTestKeyRequestPayload(key_type=type))
+    return client.services.regenerate_test_key(resource_group, name,
+                                               models.RegenerateTestKeyRequestPayload(key_type=type))
 
 
 def app_append_persistent_storage(cmd, client, resource_group, service, name,
@@ -300,7 +303,8 @@ def app_create(cmd, client, resource_group, service, name,
                env=None,
                enable_persistent_storage=None,
                assign_identity=None,
-               persistent_storage=None):
+               persistent_storage=None,
+               loaded_public_certificate_file=None):
     cpu = validate_cpu(cpu)
     memory = validate_memory(memory)
     apps = _get_all_apps(client, resource_group, service)
@@ -351,6 +355,22 @@ def app_create(cmd, client, resource_group, service, name,
                         custom_persistent_disk_properties=custom_persistent_disk_properties))
         properties.custom_persistent_disks = custom_persistent_disks
 
+    if loaded_public_certificate_file is not None:
+        data = get_file_json(loaded_public_certificate_file)
+        if data:
+            if not data.get('loadedCertificates'):
+                raise FileOperationError("loadedCertificates must be provided in the json file")
+            loaded_certificates = []
+            for item in data['loadedCertificates']:
+                invalidProperties = not item.get('certificateName') or not item.get('loadTrustStore')
+                if invalidProperties:
+                    raise FileOperationError("certificateName, loadTrustStore must be provided in the json file")
+                certificate_resource = client.certificates.get(resource_group, service, item['certificateName'])
+                loaded_certificates.append(models_20210901preview.
+                                           LoadedCertificate(resource_id=certificate_resource.id,
+                                                             load_trust_store=item['loadTrustStore']))
+            properties.loaded_certificates = loaded_certificates
+
     app_resource = models_20210901preview.AppResource()
     app_resource.properties = properties
     app_resource.location = resource.location
@@ -365,7 +385,8 @@ def app_create(cmd, client, resource_group, service, name,
     # create default deployment
     logger.warning(
         "[2/4] Creating default deployment with name '{}'".format(DEFAULT_DEPLOYMENT_NAME))
-    default_deployment_resource = _default_deployment_resource_builder(cpu, memory, env, jvm_options, runtime_version, instance_count)
+    default_deployment_resource = _default_deployment_resource_builder(cpu, memory, env, jvm_options, runtime_version,
+                                                                       instance_count)
     poller = client.deployments.begin_create_or_update(resource_group,
                                                        service,
                                                        name,
@@ -434,7 +455,8 @@ def app_update(cmd, client, resource_group, service, name,
                enable_persistent_storage=None,
                https_only=None,
                enable_end_to_end_tls=None,
-               persistent_storage=None):
+               persistent_storage=None,
+               loaded_public_certificate_file=None):
     _check_active_deployment_exist(client, resource_group, service, name)
     resource = client.services.get(resource_group, service)
     location = resource.location
@@ -475,6 +497,21 @@ def app_update(cmd, client, resource_group, service, name,
                         storage_id=storage_resource.id,
                         custom_persistent_disk_properties=custom_persistent_disk_properties))
         properties.custom_persistent_disks = custom_persistent_disks
+    if loaded_public_certificate_file is not None:
+        data = get_file_json(loaded_public_certificate_file)
+        if data:
+            if not data.get('loadedCertificates'):
+                raise CLIError("loadedCertificates must be provided in the json file")
+            loaded_certificates = []
+            for item in data['loadedCertificates']:
+                invalidProperties = not item.get('certificateName') or not item.get('loadTrustStore')
+                if invalidProperties:
+                    raise CLIError("certificateName, loadTrustStore must be provided in the json file")
+                certificate_resource = client.certificates.get(resource_group, service, item['certificateName'])
+                loaded_certificates.append(models_20210901preview.
+                                           LoadedCertificate(resource_id=certificate_resource.id,
+                                                             load_trust_store=item['loadTrustStore']))
+            properties.loaded_certificates = loaded_certificates
 
     app_resource = models_20210901preview.AppResource()
     app_resource.properties = properties
@@ -502,7 +539,7 @@ def app_update(cmd, client, resource_group, service, name,
         environment_variables=env,
         jvm_options=jvm_options,
         net_core_main_entry_path=main_entry,
-        runtime_version=runtime_version,)
+        runtime_version=runtime_version)
     deployment_settings.cpu = None
     deployment_settings.memory_in_gb = None
     properties = models_20210901preview.DeploymentResourceProperties(
@@ -854,6 +891,36 @@ def app_unset_deployment(cmd, client, resource_group, service, name):
     app_resource.location = location
 
     return client.apps.begin_update(resource_group, service, name, app_resource)
+
+
+def app_append_loaded_public_certificate(cmd, client, resource_group, service, name, certificate_name, load_trust_store):
+    app_resource = client.apps.get(resource_group, service, name)
+    certificate_resource = client.certificates.get(resource_group, service, certificate_name)
+    certificate_resource_id = certificate_resource.id
+
+    loaded_certificates = []
+    if app_resource.properties.loaded_certificates:
+        for loaded_certificate in app_resource.properties.loaded_certificates:
+            loaded_certificates.append(loaded_certificate)
+
+    for loaded_certificate in loaded_certificates:
+        if loaded_certificate.resource_id == certificate_resource.id:
+            raise ClientRequestError("This certificate has already been loaded.")
+
+    loaded_certificates.append(models_20210901preview.
+                               LoadedCertificate(resource_id=certificate_resource_id,
+                                                 load_trust_store=load_trust_store))
+
+    app_resource.properties.loaded_certificates = loaded_certificates
+    logger.warning("[1/1] updating app '{}'".format(name))
+
+    poller = client.apps.begin_update(
+        resource_group, service, name, app_resource)
+    while poller.done() is False:
+        sleep(APP_CREATE_OR_UPDATE_SLEEP_INTERVAL)
+
+    app_updated = client.apps.get(resource_group, service, name)
+    return app_updated
 
 
 def deployment_create(cmd, client, resource_group, service, app, name,
@@ -1426,11 +1493,14 @@ def _app_deploy(client, resource_group, service, app, name, version, path, runti
     if file_type is None:
         # update means command is az spring-cloud app deploy xxx
         if update:
-            raise RequiredArgumentMissingError("One of the following arguments are required: '--artifact-path' or '--source-path'")
+            raise RequiredArgumentMissingError(
+                "One of the following arguments are required: '--artifact-path' or '--source-path'")
         # if command is az spring-cloud app deployment create xxx, create default deployment
         else:
-            logger.warning("Creating default deployment without artifact/source folder. Please specify the --artifact-path/--source-path argument explicitly if needed.")
-            default_deployment_resource = _default_deployment_resource_builder(cpu, memory, env, jvm_options, runtime_version, instance_count)
+            logger.warning(
+                "Creating default deployment without artifact/source folder. Please specify the --artifact-path/--source-path argument explicitly if needed.")
+            default_deployment_resource = _default_deployment_resource_builder(cpu, memory, env, jvm_options,
+                                                                               runtime_version, instance_count)
             return sdk_no_wait(no_wait, client.deployments.begin_create_or_update,
                                resource_group, service, app, name, default_deployment_resource)
     upload_url = None
@@ -1586,7 +1656,7 @@ def _get_app_log(url, user_name, password, format_json, exceptions):
 
         return format_line
 
-    def iter_lines(response, limit=2**20):
+    def iter_lines(response, limit=2 ** 20):
         '''
         Returns a line iterator from the response content. If no line ending was found and the buffered content size is
         larger than the limit, the buffer will be yielded directly.
@@ -1702,12 +1772,80 @@ def storage_list_persistent_storage(client, resource_group, service, name):
     return reference_apps
 
 
-def certificate_list_reference_app(cmd, resource_group, service, name):
-    client = get_mgmt_service_client(cmd.cli_ctx, AppPlatformManagementClient_20210901preview)
+def certificate_add(cmd, client, resource_group, service, name, only_public_cert=None,
+                    vault_uri=None, vault_certificate_name=None, public_certificate_file=None):
+    if vault_uri is None and public_certificate_file is None:
+        raise InvalidArgumentValueError("One of --vault-uri and --public-certificate-file should be provided")
+    if vault_uri is not None and public_certificate_file is not None:
+        raise InvalidArgumentValueError("--vault-uri and --public-certificate-file could not be provided at the same time")
+    if vault_uri is not None:
+        if vault_certificate_name is None:
+            raise InvalidArgumentValueError("--vault-certificate-name should be provided for Key Vault Certificate")
 
+    if vault_uri is not None:
+        if only_public_cert is None:
+            only_public_cert = False
+        properties = models_20210901preview.KeyVaultCertificateProperties(
+            type="KeyVaultCertificate",
+            vault_uri=vault_uri,
+            key_vault_cert_name=vault_certificate_name,
+            exclude_private_key=only_public_cert
+        )
+    else:
+        if os.path.exists(public_certificate_file):
+            try:
+                with open(public_certificate_file, 'rb') as input_file:
+                    logger.debug("attempting to read file %s as binary", public_certificate_file)
+                    content = base64.b64encode(input_file.read()).decode("utf-8")
+            except Exception:
+                raise FileOperationError('Failed to decode file {} - unknown decoding'.format(public_certificate_file))
+        else:
+            raise FileOperationError("public_certificate_file %s could not be found", public_certificate_file)
+        properties = models_20210901preview.ContentCertificateProperties(
+            type="ContentCertificate",
+            content=content
+        )
+    certificate_resource = models_20210901preview.CertificateResource(properties=properties)
+
+    def callback(pipeline_response, deserialized, headers):
+        return models_20210901preview.CertificateResource.deserialize(json.loads(pipeline_response.http_response.text()))
+
+    return client.certificates.begin_create_or_update(
+        resource_group_name=resource_group,
+        service_name=service,
+        certificate_name=name,
+        certificate_resource=certificate_resource,
+        cls=callback
+    )
+
+
+def certificate_show(cmd, client, resource_group, service, name):
+    return client.certificates.get(resource_group, service, name)
+
+
+def certificate_list(cmd, client, resource_group, service, certificate_type=None):
+    certificates = list(client.certificates.list(resource_group, service))
+    certificates_to_list = []
+    if certificate_type is None:
+        certificates_to_list = certificates
+    elif certificate_type == 'KeyVaultCertificate':
+        for certificate in certificates:
+            if certificate.properties.type == 'KeyVaultCertificate':
+                certificates_to_list.append(certificate)
+    elif certificate_type == 'ContentCertificate':
+        for certificate in certificates:
+            if certificate.properties.type == 'ContentCertificate':
+                certificates_to_list.append(certificate)
+    return certificates_to_list
+
+
+def certificate_remove(cmd, client, resource_group, service, name):
+    client.certificates.get(resource_group, service, name)
+    return client.certificates.begin_delete(resource_group, service, name)
+
+
+def certificate_list_reference_app(cmd, client, resource_group, service, name):
     apps = list(client.apps.list(resource_group, service))
-    deployments = list(
-        client.deployments.list_for_cluster(resource_group, service))
     reference_apps = []
     certificate_resource = client.certificates.get(resource_group, service, name)
     certificate_resource_id = certificate_resource.id
@@ -1716,38 +1854,7 @@ def certificate_list_reference_app(cmd, resource_group, service, name):
             if load_certificate.resource_id == certificate_resource_id:
                 reference_apps.append(app)
                 break
-    for app in reference_apps:
-        if app.properties.active_deployment_name:
-            deployment = next(
-                (x for x in deployments if x.properties.app_name == app.name))
-            app.properties.active_deployment = deployment
     return reference_apps
-
-
-def certificate_add(cmd, client, resource_group, service, name, vault_uri, vault_certificate_name):
-    properties = models.CertificateProperties(
-        vault_uri=vault_uri,
-        key_vault_cert_name=vault_certificate_name
-    )
-    certificate_resource = models.CertificateResource(properties=properties)
-    return client.certificates.begin_create_or_update(
-        resource_group_name=resource_group,
-        service_name=service,
-        certificate_name=name,
-        certificate_resource=certificate_resource)
-
-
-def certificate_show(cmd, client, resource_group, service, name):
-    return client.certificates.get(resource_group, service, name)
-
-
-def certificate_list(cmd, client, resource_group, service):
-    return client.certificates.list(resource_group, service)
-
-
-def certificate_remove(cmd, client, resource_group, service, name):
-    client.certificates.get(resource_group, service, name)
-    return client.certificates.begin_delete(resource_group, service, name)
 
 
 def domain_bind(cmd, client, resource_group, service, app,